Move validations from the routes to a separated file
